refactor: 주석 삭제 및 RowMapper 내재화, test 오류 수정
