+++ src/main/java/gift/controller/ProductController.java
@@ -26,26 +26,22 @@ public class ProductController {
         this.productService = productService;
     }
 
-    // 생성
     @PostMapping
     public ResponseEntity<ProductResponseDto> createProduct(
             @Valid @RequestBody ProductRequestDto requestDto) {
         return new ResponseEntity<>(productService.save(requestDto), HttpStatus.CREATED);
     }
 
-    // 목록 조회
     @GetMapping
     public List<ProductResponseDto> findAll() {
         return productService.findAll();
     }
 
-    // 단건 조회
     @GetMapping("/{id}")
     public ResponseEntity<ProductResponseDto> findById(@PathVariable Long id) {
         return new ResponseEntity<>(productService.findById(id), HttpStatus.OK);
     }
 
-    // 수정
     @PutMapping("/{id}")
     public ResponseEntity<ProductResponseDto> update(
             @PathVariable Long id,
@@ -56,7 +52,6 @@ public class ProductController {
                         requestDto.imageUrl()), HttpStatus.OK);
     }
 
-    // 삭제
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
         productService.delete(id);
