--- src/main/java/gift/repository/ProductRowMapper.java
@@ -1,19 +1,0 @@
-package gift.repository;
-
-import gift.entity.Product;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import org.springframework.jdbc.core.RowMapper;
-
-public class ProductRowMapper implements RowMapper<Product> {
-
-    @Override
-    public Product mapRow(ResultSet rs, int rowNum) throws SQLException {
-        return new Product(
-                rs.getLong("id"),
-                rs.getString("name"),
-                rs.getInt("price"),
-                rs.getString("image_url")
-        );
-    }
-}
